refactor: architecture && added: pdf reader and files for reading
