+++ informatics-text-analyzer.iml
@@ -3,7 +3,7 @@
   <component name="NewModuleRootManager" inherit-compiler-output="true">
     <exclude-output />
     <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/resources" type="java-resource" />
+      <sourceFolder url="file://$MODULE_DIR$/text-analyzer/src/resources" type="java-resource" />
     </content>
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
