+++ tasks
@@ -1,0 +1,18 @@
+Азамат 1-3
+Багат 4-6
+Юлия 7-9
+Аделя 10-12
+Даниил 13-15
+Камиль Хузин 16-18
+Камиль Вафин 19-21
+Аделя Халяпова 22-24
+Артем 25-27
+Иван 28-30
+Айдар 31-33
+Саша 34-36
+Карима 37-39
+Дария 40-42
+Булат 43-45
+Ильмир 45-46
+Аня 47
+Рашид 48
