+++ text-analyzer/src/resources/text.txt
